bf 16

+++ get_model.py
@@ -22,5 +22,6 @@ def get_model(model_id):
                                                 device_map="auto", 
                                                 trust_remote_code=True, 
                                                 offload_folder="offload", 
-                                                torch_dtype = torch.float16)
+                                                torch_dtype = torch.bfloat16
+                                                )
     return tokenizer, model
